backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
google-cloud-aiplatform==1.38.1
python-multipart==0.0.7
pydantic>=2.9
pydantic-settings==2.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.49
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
google-genai==1.30.0
openai>=1.68.0
anthropic>=0.49.0